backend/requirements.txt
fastapi==0.115.2
uvicorn[standard]==0.30.6
motor==3.6.0
pymongo==4.9.1
python-multipart==0.0.9
pydantic==2.9.2
celery==5.4.0
redis==5.0.8
torch
torchvision
numpy>=1.26.4
opencv-python==4.10.0.84
pillow==10.4.0
scipy>=1.11.4
mtcnn==0.1.1
facenet-pytorch==2.5.3
tensorflow>=2.20.0
typing-extensions>=4.7.0
pytest==8.3.3
pytest-asyncio==0.24.0
httpx==0.27.2
python-dotenv==1.0.1
















